moving fetching to each row for better performance and to prevent double fetches
