+++ app/components/scanner/drawer/generic-item-row.tsx
@@ -1,5 +1,8 @@
-// components/scanner/generic-item-row.tsx
+import { useCallback, useEffect, useRef } from "react";
 import { motion } from "framer-motion";
+import { useSetAtom } from "jotai";
+import type { ScanListItem } from "~/atoms/qr-scanner";
+import { updateScannedItemAtom } from "~/atoms/qr-scanner";
 import { Button } from "~/components/shared/button";
 import { Td } from "~/components/table";
 import { tw } from "~/utils/tw";
@@ -7,16 +10,15 @@ import { tw } from "~/utils/tw";
 // Type for the row props
 type GenericItemRowProps<T> = {
   qrId: string;
-  item: T | null;
+  item: ScanListItem | undefined;
   onRemove: (qrId: string) => void;
   renderItem: (item: T) => React.ReactNode;
   renderLoading: (qrId: string, error?: string) => React.ReactNode;
-  hasError?: boolean;
-  error?: string;
 };
 
 /**
  * Generic component for rendering a row in the scanned items table
+ * With self-contained fetch functionality
  */
 export function GenericItemRow<T>({
   qrId,
@@ -24,14 +26,87 @@ export function GenericItemRow<T>({
   onRemove,
   renderItem,
   renderLoading,
-  hasError,
-  error,
 }: GenericItemRowProps<T>) {
+  const setItem = useSetAtom(updateScannedItemAtom);
+  const hasFetched = useRef(false);
+
+  /**
+   * Fetch item data for this specific row
+   */
+  const fetchItem = useCallback(async () => {
+    // Skip if we already have data or if we've already tried to fetch
+    if (hasFetched.current || (item && (item.data || item.error))) {
+      return;
+    }
+
+    // Mark as fetched to prevent duplicate fetches
+    hasFetched.current = true;
+
+    try {
+      const request = await fetch(`/api/get-scanned-item/${qrId}`);
+      const response = await request.json();
+
+      /** If the server returns an error, add it to the item and return */
+      if (response.error) {
+        setItem({
+          qrId,
+          item: { error: response.error.message },
+        });
+        return;
+      }
+
+      const qr = response.qr;
+
+      /** Determine item type (asset or kit) and update accordingly */
+      if (qr && qr.type === "asset") {
+        const itemWithType: ScanListItem = {
+          data: qr.asset,
+          type: "asset",
+        };
+
+        if (itemWithType.data) {
+          setItem({
+            qrId,
+            item: itemWithType,
+          });
+        }
+      } else if (qr && qr.type === "kit") {
+        const itemWithType: ScanListItem = {
+          data: qr.kit,
+          type: "kit",
+        };
+
+        if (itemWithType.data) {
+          setItem({
+            qrId,
+            item: itemWithType,
+          });
+        }
+      }
+    } catch (error) {
+      setItem({
+        qrId,
+        item: { error: "Failed to fetch item" },
+      });
+    }
+  }, [qrId, item, setItem]);
+
+  /** Fetch data when the component mounts */
+  useEffect(() => {
+    void fetchItem();
+  }, [fetchItem]);
+
+  // Determine if we should show the item or loading state
+  // Only show the item if we have both data and type (complete item)
+  const shouldShowItem = item && item.data && item.type && !item.error;
+
   return (
     <Tr>
       <Td className="w-full p-0 md:p-0">
         <div className="flex items-center justify-between gap-3 p-4 md:px-6">
-          {!item || hasError ? renderLoading(qrId, error) : renderItem(item)}
+          {shouldShowItem
+            ? renderItem(item.data as unknown as T)
+            : renderLoading(qrId, item?.error)}
         </div>
       </Td>
       <Td>
