+++ app/components/scanner/drawer/uses/add-assets-to-booking-drawer.tsx
@@ -199,9 +199,7 @@ export default function AddAssetsToBookingDrawer({
     <GenericItemRow
       key={qrId}
       qrId={qrId}
-      item={item?.data}
-      hasError={!!item?.error}
-      error={item?.error}
+      item={item}
       onRemove={removeItem}
       renderLoading={(qrId, error) => (
         <DefaultLoadingState qrId={qrId} error={error} />
